feat(error): error codes, and printing function
Add an enum for error codes and a function to print a string based on
the code. This is so the error codes are nicely managed without having
to explicitly set a value (and change the values if you want to make an
entry in between codes), and automatically keeps a value at the end for
easy bounds checking.